Changed whole numbers to float values

+++ Calculator.java
@@ -5,15 +5,15 @@ public class Calculator {
         Scanner scan = new Scanner(System.in);
 
         System.out.println("Please enter the first whole number");
-        int num1 = scan.nextInt();
+        float num1 = scan.nextInt();
 
         System.out.println("Please enter the second whole number");
-        int num2 = scan.nextInt();
+        float num2 = scan.nextInt();
 
         System.out.println("Would you like to add or subtract these numbers? Enter 1 for add, or 2 for subtract, 3 for multiply, and 4 for divide.");
-        int choice = scan.nextInt();
+        float choice = scan.nextInt();
 
-        int sum = 0;
+        float sum = 0;
 
         if (choice == 1) {
             sum = num1 + num2;
